Add win32 support
Replace Make-based compilation of the nif to using rebar's port
compiler.

+++ c_src/enacl_nif.c
@@ -1515,7 +1515,7 @@ static
 crypto_generichash_state *align64(void *ptr){
   if((unsigned long)ptr % 64 == 0)
     return ptr;
-  return ptr + (64 - ((unsigned long)ptr % 64));
+  return (unsigned long)ptr + (64 - ((unsigned long)ptr % 64));
 }
 
 static
